refactor: user -> member 패키지명 수정
