+++ src/main/java/finalmission/member/domain/vo/Password.java
@@ -1,4 +1,4 @@
-package finalmission.user.domain.vo;
+package finalmission.member.domain.vo;
 
 import jakarta.persistence.Embeddable;
