+++ src/main/java/finalmission/member/domain/vo/Role.java
@@ -1,4 +1,4 @@
-package finalmission.user.domain.vo;
+package finalmission.member.domain.vo;
 
 import java.util.Arrays;
