+++ src/main/java/finalmission/reservation/domain/Reservation.java
@@ -1,6 +1,6 @@
 package finalmission.reservation.domain;
 
-import finalmission.user.domain.Member;
+import finalmission.member.domain.Member;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
